requirements.txt
aiogram==2.25.1
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.0.1